standardize random number seed parameters in c++ source code to use unsigned long long type uniformly

+++ src/SGC4Lattice.h
@@ -127,7 +127,7 @@ std::vector<double> SGC4Lattice(
     int threads,
     int boot = 399,
     double base = 2,
-    unsigned int seed = 42,
+    unsigned long long seed = 42,
     bool symbolize = true,
     bool normalize = false,
     bool progressbar = true
